test(foundation): S6 Inc 1-4 checkpoint — H1-loom + fault-injection harness + 2 drills
Partial S6 (Batteries D + H1-loom) — XL, /review-impl'd on spec+plan. The
conformance cases + CI + final docs land in the completing commit (Inc 5-6).

Inc 1 — H1-loom: extracted the tokio-free CircuitBreaker sync core into a new
leaf crate `breaker-core` (loom can't build inside dp-kernel — tokio's transitive
deps break under --cfg loom). dp-kernel::resilience re-exports CircuitBreaker so
every consumer's path is unchanged. loom exhaustively checks the gate()->record()
split-lock interleavings + a should_panic bite (loom catches an unsynchronized
counter). RUSTFLAGS="--cfg loom" cargo test -p breaker-core: 2 pass, <1ms model.

Inc 2 — injection harness: toxiproxy-foundation compose service (admin :8474,
pg_proxy/redis_proxy listen ports pre-published) + scripts/chaos/toxic.sh driver
(latency/timeout/down/up/reset + selftest). Selftest: latency toxic added 2584ms
on the real PG path.

Inc 3 — fault-pg-down.sh (graceful-degradation, emit path): deterministic bracket
down -> emit fails fast (294ms, EOF) -> 0 partial rows (transactional) -> up ->
fresh emit + C3 verify clean. NOT a convergence claim (interrupted non-idempotent
emit can't be completed — review-impl #1).

Inc 4 — fault-redis-partition.sh (convergence, publish path) + bash history
checker: seed -> publisher drains outbox->Redis -> down redis mid-drain (pending,
attempts>0, no dead-letter) -> heal -> drained -> C3 + history (no-loss/no-dup,
events==XLEN). --bite XDELs a delivered stream entry: the history checker flags
the loss while the outbox still shows all published (it sees what post-quiesce
drift misses). Deterministic via BATCH_SIZE=5 mid-drain partition.

VERIFY (live): all four increments green on the foundation-dev stack. Windows:
proxy host ports avoid the excluded range (FOUNDATION_PG_PROXY_PORT=55601);
runner-green for loom is per-PR-able (CPU only).
